+++ src/project_remote/repo.rs
@@ -2,7 +2,7 @@ use std::fs;
 use std::path::Path;
 
 use git2::{build::CheckoutBuilder, ErrorCode, FetchOptions, Oid, Remote, Repository};
-use orion_error::{ToStructError, UvsFrom};
+use orion_error::{conversion::ToStructError, UvsFrom};
 use semver::Version;
 use wp_error::run_error::RunResult;
 use wp_error::RunReason;